+++ src/DbConnectionExtensions.Execute.cs
@@ -6,7 +6,7 @@ using System.Data.Common;
 /// <summary>
 /// Provides extension members for database connections.
 /// </summary>
-public static partial class ConnectionExtensions {
+public static partial class DbConnectionExtensions {
 
 	/// <summary>
 	/// Executes a parameterized SQL statement.
@@ -16,7 +16,7 @@ public static partial class ConnectionExtensions {
 	/// <param name="parameters">The parameters of the SQL query.</param>
 	/// <param name="options">The command options.</param>
 	/// <returns>The number of rows affected.</returns>
-	public static int Execute(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null) {
+	public static int Execute(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null) {
 		if (connection.State == ConnectionState.Closed) connection.Open();
 		using var dbCommand = CreateCommand(connection, text, parameters, options);
 		return dbCommand.ExecuteNonQuery();
@@ -31,7 +31,7 @@ public static partial class ConnectionExtensions {
 	/// <param name="options">The command options.</param>
 	/// <param name="cancellationToken">The token to cancel the operation.</param>
 	/// <returns>The number of rows affected.</returns>
-	public static async Task<int> ExecuteAsync(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null, CancellationToken cancellationToken = default) {
+	public static async Task<int> ExecuteAsync(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null, CancellationToken cancellationToken = default) {
 		if (connection.State == ConnectionState.Closed) await ((DbConnection) connection).OpenAsync(cancellationToken);
 		using var dbCommand = (DbCommand) CreateCommand(connection, text, parameters, options);
 		return await dbCommand.ExecuteNonQueryAsync(cancellationToken);
@@ -45,7 +45,7 @@ public static partial class ConnectionExtensions {
 	/// <param name="parameters">The parameters of the SQL query.</param>
 	/// <param name="options">The command options.</param>
 	/// <returns>The data reader that can be used to access the results.</returns>
-	public static IDataReader ExecuteReader(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null) {
+	public static IDataReader ExecuteReader(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null) {
 		if (connection.State == ConnectionState.Closed) connection.Open();
 		using var dbCommand = CreateCommand(connection, text, parameters, options);
 		return dbCommand.ExecuteReader();
@@ -60,7 +60,7 @@ public static partial class ConnectionExtensions {
 	/// <param name="options">The command options.</param>
 	/// <param name="cancellationToken">The token to cancel the operation.</param>
 	/// <returns>The data reader that can be used to access the results.</returns>
-	public static async Task<IDataReader> ExecuteReaderAsync(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null, CancellationToken cancellationToken = default) {
+	public static async Task<IDataReader> ExecuteReaderAsync(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null, CancellationToken cancellationToken = default) {
 		if (connection.State == ConnectionState.Closed) await ((DbConnection) connection).OpenAsync(cancellationToken);
 		using var dbCommand = (DbCommand) CreateCommand(connection, text, parameters, options);
 		return await dbCommand.ExecuteReaderAsync(cancellationToken);
@@ -74,7 +74,7 @@ public static partial class ConnectionExtensions {
 	/// <param name="parameters">The parameters of the SQL query.</param>
 	/// <param name="options">The command options.</param>
 	/// <returns>The first column of the first row.</returns>
-	public static object? ExecuteScalar(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null) =>
+	public static object? ExecuteScalar(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null) =>
 		ExecuteScalar<object>(connection, text, parameters, options);
 
 	/// <summary>
@@ -86,7 +86,7 @@ public static partial class ConnectionExtensions {
 	/// <param name="options">The command options.</param>
 	/// <param name="cancellationToken">The token to cancel the operation.</param>
 	/// <returns>The first column of the first row.</returns>
-	public static async Task<object?> ExecuteScalarAsync(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null, CancellationToken cancellationToken = default) =>
+	public static async Task<object?> ExecuteScalarAsync(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null, CancellationToken cancellationToken = default) =>
 		await ExecuteScalarAsync<object>(connection, text, parameters, options, cancellationToken);
 
 	/// <summary>
@@ -98,11 +98,11 @@ public static partial class ConnectionExtensions {
 	/// <param name="parameters">The parameters of the SQL query.</param>
 	/// <param name="options">The command options.</param>
 	/// <returns>The first column of the first row.</returns>
-	public static T? ExecuteScalar<T>(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null) {
+	public static T? ExecuteScalar<T>(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null) {
 		if (connection.State == ConnectionState.Closed) connection.Open();
 		using var dbCommand = CreateCommand(connection, text, parameters, options);
 		var value = dbCommand.ExecuteScalar();
-		return value is null || value is DBNull ? default : (T?) Mapper.ChangeType(value, typeof(T));
+		return value is null || value is DBNull ? default : (T?) SqlMapper.ChangeType(value, typeof(T));
 	}
 
 	/// <summary>
@@ -115,10 +115,10 @@ public static partial class ConnectionExtensions {
 	/// <param name="options">The command options.</param>
 	/// <param name="cancellationToken">The token to cancel the operation.</param>
 	/// <returns>The first column of the first row.</returns>
-	public static async Task<T?> ExecuteScalarAsync<T>(this IDbConnection connection, string text, ParameterCollection? parameters = null, CommandOptions? options = null, CancellationToken cancellationToken = default) {
+	public static async Task<T?> ExecuteScalarAsync<T>(this IDbConnection connection, string text, SqlParameterCollection? parameters = null, SqlCommandOptions? options = null, CancellationToken cancellationToken = default) {
 		if (connection.State == ConnectionState.Closed) await ((DbConnection) connection).OpenAsync(cancellationToken);
 		using var dbCommand = (DbCommand) CreateCommand(connection, text, parameters, options);
 		var value = await dbCommand.ExecuteScalarAsync(cancellationToken);
-		return value is null || value is DBNull ? default : (T?) Mapper.ChangeType(value, typeof(T));
+		return value is null || value is DBNull ? default : (T?) SqlMapper.ChangeType(value, typeof(T));
 	}
 }