+++ test/DbConnectionExtensions.Query.Tests.cs
@@ -1,30 +1,16 @@
 namespace Belin.Sql;
 
 using Belin.Sql.Fixtures;
-using System.Data.SQLite;
 
 /// <summary>
-/// Tests the features of the <see cref="ConnectionExtensions"/> class.
+/// Tests the features of the <see cref="DbConnectionExtensions"/> class.
 /// </summary>
-/// <param name="testContext">The test context.</param>
-[TestClass]
-public sealed class ConnectionExtensionsQueryTests(TestContext testContext) {
-
-	/// <summary>
-	/// The connection to the data source.
-	/// </summary>
-	private SQLiteConnection connection = default!;
-
-	[TestInitialize]
-	public void TestInitialize() => connection = That.CreateInMemoryDatabase();
-
-	[TestCleanup]
-	public void TestCleanup() => connection.Close();
+public sealed partial class DbConnectionExtensionsTests {
 
 	[TestMethod]
 	public async Task Query() {
 		var sql = "SELECT * FROM Characters WHERE Gender = @Gender ORDER BY FullName";
-		var parameters = new ParameterCollection(("Gender", CharacterGender.Elf.ToString()));
+		var parameters = new SqlParameterCollection(("Gender", CharacterGender.Elf.ToString()));
 
 		var records = connection.Query<Character>(sql, parameters).AsList();
 		HasCount(3, records);
@@ -44,7 +30,7 @@ public sealed class ConnectionExtensionsQueryTests(TestContext testContext) {
 	[TestMethod]
 	public async Task QueryFirst() {
 		var sql = "SELECT * FROM Characters WHERE FullName = @FullName";
-		var parameters = new ParameterCollection(("FullName", "Sauron"));
+		var parameters = new SqlParameterCollection(("FullName", "Sauron"));
 
 		var record = connection.QueryFirst<Character>(sql, parameters);
 		AreEqual("Sauron", record.FirstName);
@@ -58,7 +44,7 @@ public sealed class ConnectionExtensionsQueryTests(TestContext testContext) {
 	[TestMethod]
 	public async Task QuerySingle() {
 		var sql = "SELECT * FROM Characters WHERE FullName = @FullName";
-		var parameters = new ParameterCollection(("FullName", "Saruman"));
+		var parameters = new SqlParameterCollection(("FullName", "Saruman"));
 
 		var record = connection.QuerySingle<Character>(sql, parameters);
 		AreEqual("Saruman", record.FirstName);